feat: set compact version as default
