--- src/frame/utils.ts
@@ -1,8 +1,0 @@
-import type { FrameState } from './types'
-
-export const defaultState = (): FrameState => ({
-  delta: 0.0,
-  timestamp: 0.0,
-  isRunning: false,
-  isPaused: false,
-})
