+++ test/frame.test.ts
@@ -1,10 +1,10 @@
 // @vitest-environment jsdom
 
-import { createFrame } from '@/frame'
+import { createFrame } from '@'
 import { describe, test } from 'vitest'
 
-describe('Frame Manager', () => {
-  test('should create a frame manager with default phases', () => {
+describe('Frame Manager - Compact', () => {
+  test('should create a compact frame manager', () => {
     return new Promise<void>((resolve, reject) => {
       const frame = createFrame()
 
@@ -14,51 +14,31 @@ describe('Frame Manager', () => {
       results.push(frame.state.delta === 0 ? true : false)
       results.push(frame.state.timestamp === 0 ? true : false)
       results.push(frame.state.isRunning === false ? true : false)
-      results.push(frame.state.isPaused === false ? true : false)
 
       return results.every((v) => v === true) ? resolve() : reject()
     })
   })
 
-  test('should run phases in the correct order: read -> update -> render', () => {
+  test('should run phases in the correct order: 0 -> 1 -> 2', () => {
     return new Promise<void>((resolve, reject) => {
       const frame = createFrame()
 
       const phases: string[] = []
 
-      frame.render(() => phases.push('render'))
-      frame.update(() => phases.push('update'))
-      frame.read(() => phases.push('read'))
-      frame.render(() => phases.push('render'))
-      frame.read(() => phases.push('read'))
-      frame.update(() => phases.push('update'))
-
-      const order = ['read', 'read', 'update', 'update', 'render', 'render']
-
-      return phases.every((v, i) => v === order[i]) ? resolve() : reject()
-    })
-  })
-
-  test('should run custom phases in the correct order: measure -> mutate', () => {
-    return new Promise<void>((resolve, reject) => {
-      const frame = createFrame({ phases: ['measure', 'mutate'] })
-
-      const phases: string[] = []
-
-      frame.mutate(() => phases.push('mutate'))
-      frame.measure(() => phases.push('measure'))
-      frame.mutate(() => phases.push('mutate'))
-      frame.measure(() => phases.push('measure'))
-      frame.mutate(() => phases.push('mutate'))
-      frame.mutate(() => phases.push('mutate'))
+      frame.add(() => phases.push('phase 2'), { phase: 2 })
+      frame.add(() => phases.push('phase 1'), { phase: 1 })
+      frame.add(() => phases.push('phase 0'))
+      frame.add(() => phases.push('phase 2'), { phase: 2 })
+      frame.add(() => phases.push('phase 0'))
+      frame.add(() => phases.push('phase 1'), { phase: 1 })
 
       const order = [
-        'measure',
-        'measure',
-        'measure',
-        'mutate',
-        'mutate',
-        'mutate',
+        'phase 0',
+        'phase 0',
+        'phase 1',
+        'phase 1',
+        'phase 2',
+        'phase 2',
       ]
 
       return phases.every((v, i) => v === order[i]) ? resolve() : reject()
@@ -71,8 +51,8 @@ describe('Frame Manager', () => {
 
       let isTriggered = false
 
-      const onUpdate = frame.update(() => (isTriggered = true))
-      frame.render(() => frame.cancel(onUpdate))
+      const process = frame.add(() => (isTriggered = true))
+      frame.add(() => frame.delete(process), { phase: 1 })
 
       return !isTriggered ? resolve() : reject()
     })
@@ -84,11 +64,11 @@ describe('Frame Manager', () => {
 
       let index = 0
 
-      frame.update(() => index++, { loop: true })
-      frame.render(
+      frame.add(() => index++, { loop: true })
+      frame.add(
         () => {
           if (index >= 3) {
-            frame.clear()
+            frame.delete()
             resolve()
           }
         },
@@ -103,11 +83,11 @@ describe('Frame Manager', () => {
 
       let index = 0
 
-      frame.update(() => {
+      frame.add(() => {
         index++
-        frame.update(() => index++, { schedule: false })
+        frame.add(() => index++, { schedule: false })
       })
-      frame.render(() => (index === 2 ? resolve() : reject()))
+      frame.add(() => (index === 2 ? resolve() : reject()), { phase: 1 })
     })
   })
 })
